+++ simgui.json
@@ -3,5 +3,8 @@
     "types": {
       "/FMSInfo": "FMSInfo"
     }
+  },
+  "NetworkTables Info": {
+    "visible": true
   }
 }
